# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Manager API Contract
  version: 0.1.0
  description: |
    Jawny kontrakt API dla endpointów objętych etapem porządkowania integracji frontend ↔ backend.
servers:
  - url: /api
components:
  schemas:
    ApiSuccessEnvelope:
      type: object
      required: [payload]
      properties:
        payload: {}
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
            code:
              type: string
            details:
              type: object
              additionalProperties: true
...
    LoanMutationResult:
      type: object
      required: [message]
      properties:
        id:
          type: integer
        message:
          type: string
    LoanSchedule:
      type: object
      required: [loan, baseline, simulation, actual_metrics, simulated_metrics, overpayments_list]
      properties:
        loan:
          type: object
          additionalProperties: true
        baseline:
          type: object
          additionalProperties: true
        simulation:
          type: object
          additionalProperties: true
        actual_metrics:
          type: object
          additionalProperties: true
        simulated_metrics:
          type: object
          additionalProperties: true
        overpayments_list:
          type: array
          items:
            type: object
            additionalProperties: true
paths:
...
  /loans/:
    get:
      summary: List loans
      responses:
        '200':
          description: Loans list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEnvelope'
    post:
      summary: Create loan
      responses:
        '201':
          description: Loan created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessEnvelope'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/LoanMutationResult'
        '400':
          description: Missing field
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /loans/{loan_id}/schedule:
    get:
      summary: Get loan schedule
      parameters:
        - in: path
          name: loan_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Loan schedule
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessEnvelope'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/LoanSchedule'
        '404':
          description: Loan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'